refactor(efi): own the handle, event, and image tables in the efi module
The Boot Services tables and their entry types move to `efi::tables`, with
`tables()` and `with_tables_mut()` as the only access path. `EfiState` is
renamed `Tables` since that is all it holds now.
